requirements.txt
aws-cdk.core==1.109.0
aws_cdk.aws_lambda
aws_cdk.aws_stepfunctions
aws_cdk.aws_stepfunctions_tasks